add groupId in petitions table

+++ backend/app/Http/Controllers/API/PetitionController.php
@@ -8,7 +8,7 @@ use App\Http\Responses\ErrorResponse;
 use App\Http\Responses\OkResponse;
 use App\Models\Petition;
 use App\Models\User;
-use Illuminate\Support\Facades\Storage;
+use ErrorException;
 use Illuminate\Support\Facades\Validator;
 
 class PetitionController extends Controller
@@ -26,7 +26,7 @@ class PetitionController extends Controller
         if (empty($petitionId)) {
             return new ErrorResponse(400, 'Invalid params');
         }
-        return new OkResponse(Petition::getPetitions([$petitionId], $withOwner = true, [], true, $request->userId, true));
+        return new OkResponse(Petition::getPetitions([$petitionId], $withOwner = true, [], true, $request->userId, true, false, false));
     }
 
     public function store(SignRequest $request)
@@ -53,6 +53,9 @@ class PetitionController extends Controller
                 $text = (string)$request->text;
                 $needSignatures = (int)$request->need_signatures;
                 $directedTo = (string)$request->directed_to;
+                $photo = null;
+                $mobilePhoto = null;
+                $webPhoto = null;
 
                 $files = [];
                 $rules = [];
@@ -79,27 +82,38 @@ class PetitionController extends Controller
                 $title = Petition::filterString($title);
                 $text = Petition::filterString($text);
                 $directedTo = Petition::filterString($directedTo);
-                if (
-                    empty($title) || empty($text) || empty($needSignatures) || ((empty($mobilePhoto) || empty($webPhoto)) && empty($photo))
-                ) {
+                if (empty($title) || empty($text) || empty($needSignatures)) {
                     return new ErrorResponse(400, 'Invalid params');
                 }
-                if (empty($mobilePhoto) || empty($webPhoto)) {
+
+                if (mb_strlen($title) === 0 || mb_strlen($title) > 150 ||
+                    mb_strlen($text) === 0 || mb_strlen($text) > 3000 ||
+                    $needSignatures === 0 || $needSignatures > 10000000 ||
+                    mb_strlen($directedTo) > 255
+                ) {
+                    return new ErrorResponse(400, 'Превышены ограничения');
+                }
+
+                if (!is_null($mobilePhoto)) {
+                    $mobilePhotoSize = getimagesize($mobilePhoto);
+                    if ($mobilePhotoSize[0] < 100 || $mobilePhotoSize[1]) {
+                        return new ErrorResponse(400, 'Слишком маленькое изображение');
+                    }
+                }
+
+                if (!is_null($webPhoto)) {
+                    $webPhotoSize = getimagesize($webPhoto);
+                    if ($webPhotoSize[0] < 100 || $webPhotoSize[1]) {
+                        return new ErrorResponse(400, 'Слишком маленькое изображение');
+                    }
+                }
+
+                if ((is_null($mobilePhoto) || is_null($webPhoto)) && !is_null($photo)) {
                     $mobilePhoto = $photo;
                     $webPhoto = $photo;
                 }
 
-                if (mb_strlen($title) === 0 || mb_strlen($title) > 150 || mb_strlen($text) === 0 || mb_strlen($text) > 3000 || $needSignatures === 0 || $needSignatures > 10000000) {
-                    return new ErrorResponse(400, 'Превышены ограничения');
-                }
-
-                $mobilePhotoSize = getimagesize($mobilePhoto);
-                $webPhotoSize = getimagesize($webPhoto);
-                if ($mobilePhotoSize[0] < 100 || $mobilePhotoSize[1] < 100 || $webPhotoSize[0] < 100 || $webPhotoSize[1] < 100) {
-                    return new ErrorResponse(400, 'Слишком маленькое изображение');
-                }
-
-                return new OkResponse(Petition::createPetition($request, $title, $text, $needSignatures, $directedTo, $mobilePhoto, $webPhoto, $request->userId));
+                return new OkResponse(Petition::createPetition($request, $title, $text, $needSignatures, $directedTo, $mobilePhoto, $webPhoto));
         }
 
         return $this->getPetitions($request, $type, $offset, $petitionId, $friendIds);
@@ -124,10 +138,15 @@ class PetitionController extends Controller
         if ($petition['owner_id'] !== $request->userId) {
             return new ErrorResponse(403, 'Access denied');
         }
-        $mobilePhotoUrl = explode(config('app.server_url'), $petition->mobile_photo_url)[1];
-        $webPhotoUrl = explode(config('app.server_url'), $petition->web_photo_url)[1];
-        unlink(base_path() . '/storage/app/public/' . $mobilePhotoUrl);
-        unlink(base_path() . '/storage/app/public/' . $webPhotoUrl);
+
+        if (!is_null($petition->mobile_photo_url)) {
+            $mobilePhotoUrl = explode(config('app.server_url'), $petition->mobile_photo_url)[1];
+            unlink(base_path() . '/storage/app/public/' . $mobilePhotoUrl);
+        }
+        if (!is_null($petition->web_photo_url)) {
+            $webPhotoUrl = explode(config('app.server_url'), $petition->web_photo_url)[1];
+            unlink(base_path() . '/storage/app/public/' . $webPhotoUrl);
+        }
         $petition->delete();
         return new OkResponse(true);
     }
@@ -168,78 +187,80 @@ class PetitionController extends Controller
         if ($request->directed_to) {
             $data['directed_to'] = Petition::filterString((string)$request->directed_to);
         }
-        if (!is_null($request->images)) {
-            if (!$request->hasFile('file1') && !$request->hasFile('file2') && !$request->hasFile('file')) {
-                return new ErrorResponse(400, 'Invalid params');
-            } else if ($request->hasFile('file')) {
-                $photo = $request->file('file');
-                $files['file'] = $photo;
-                $rules['file'] = 'required|image|mimes:png,jpeg,jpg|max:10240'; // 10mb
-                $validator = Validator::make($files, $rules);
-                if ($validator->fails()) {
-                    return new ErrorResponse(400, 'Недействительное изображение');
-                }
-                $photoSize = getimagesize($photo);
-                if ($photoSize[0] < 100 || $photoSize[1] < 100) {
-                    return new ErrorResponse(400, 'Слишком маленькое изображение');
-                }
-                $saveData = Petition::saveImages($photo, $photo);
-                $data['mobile_photo_url'] = config('app.server_url') . 'static/' . $saveData['name'] . '_mobile.png';
-                $data['web_photo_url'] = config('app.server_url') . 'static/' . $saveData['name'] . '_web.png';
-                $mobilePhotoUrl = explode(config('app.server_url'), $petition['mobile_photo_url'])[1];
-                $webPhotoUrl = explode(config('app.server_url'), $petition['web_photo_url'])[1];
-                unlink(base_path() . '/storage/app/public/' . $mobilePhotoUrl);
-                unlink(base_path() . '/storage/app/public/' . $webPhotoUrl);
-
-            } else if ($request->hasFile('file1') || $request->hasFile('file2')) {
-                $mobilePhoto = null;
-                $webPhoto = null;
-                if ($request->hasFile('file1')) {
-                    $mobilePhoto = $request->file('file1');
-                    $files['file1'] = $mobilePhoto;
-                    $rules['file1'] = 'required|image|mimes:png,jpeg,jpg|max:10240'; // 10mb
-                }
-                if ($request->hasFile('file2')) {
-                    $webPhoto = $request->file('file2');
-                    $files['file2'] = $webPhoto;
-                    $rules['file2'] = 'required|image|mimes:png,jpeg,jpg|max:10240'; // 10mb
-                }
-                $validator = Validator::make($files, $rules);
-                if ($validator->fails()) {
-                    return new ErrorResponse(400, 'Недействительные изображения');
-                }
-
-                if ($mobilePhoto) {
-                    $mobilePhotoSize = getimagesize($mobilePhoto);
-                    if ($mobilePhotoSize[0] < 100 || $mobilePhotoSize[1] < 100) {
-                        return new ErrorResponse(400, 'Слишком маленькое изображение');
-                    }
-                }
-                if ($webPhoto) {
-                    $webPhotoSize = getimagesize($webPhoto);
-                    if ($webPhotoSize[0] < 100 || $webPhotoSize[1] < 100) {
-                        return new ErrorResponse(400, 'Слишком маленькое изображение');
-                    }
-                }
-
-                $saveData = Petition::saveImages($mobilePhoto, $webPhoto);
-                if ($saveData['mobilePhoto']) {
-                    $data['mobile_photo_url'] = config('app.server_url') . 'static/' . $saveData['name'] . '_mobile.png';
-                    $mobilePhotoUrl = explode(config('app.server_url'), $petition['mobile_photo_url'])[1];
-                    unlink(base_path() . '/storage/app/public/' . $mobilePhotoUrl);
-                }
-                if ($saveData['webPhoto']) {
-                    $data['web_photo_url'] = config('app.server_url') . 'static/' . $saveData['name'] . '_web.png';
-                    $webPhotoUrl = explode(config('app.server_url'), $petition['web_photo_url'])[1];
-                    unlink(base_path() . '/storage/app/public/' . $webPhotoUrl);
-                }
-            } else {
-                return new ErrorResponse(400, 'Invalid params');
-            }
-        }
         if (!is_null($request->completed)) {
             $data['completed'] = (bool)$request->completed;
         }
+
+
+        $photo = null;
+        $mobilePhoto = null;
+        $webPhoto = null;
+
+        $files = [];
+        $rules = [];
+        if ($request->hasFile('file1')) {
+            $mobilePhoto = $request->file('file1');
+            $files['file1'] = $mobilePhoto;
+            $rules['file1'] = 'required|image|mimes:png,jpeg,jpg|max:10240'; // 10mb
+        }
+        if ($request->hasFile('file2')) {
+            $webPhoto = $request->file('file2');
+            $files['file2'] = $webPhoto;
+            $rules['file2'] = 'required|image|mimes:png,jpeg,jpg|max:10240'; // 10mb
+        }
+        if ($request->hasFile('file')) {
+            $photo = $request->file('file');
+            $files['file'] = $photo;
+            $rules['file'] = 'required|image|mimes:png,jpeg,jpg|max:10240'; // 10mb
+        }
+        $validator = Validator::make($files, $rules);
+        if ($validator->fails()) {
+            return new ErrorResponse(400, 'Недействительные изображения');
+        }
+
+        if (!is_null($mobilePhoto)) {
+            $mobilePhotoSize = getimagesize($mobilePhoto);
+            if ($mobilePhotoSize[0] < 100 || $mobilePhotoSize[1] < 100) {
+                return new ErrorResponse(400, 'Слишком маленькое изображение');
+            }
+        }
+
+        if (!is_null($webPhoto)) {
+            $webPhotoSize = getimagesize($webPhoto);
+            if ($webPhotoSize[0] < 100 || $webPhotoSize[1] < 100) {
+                return new ErrorResponse(400, 'Слишком маленькое изображение');
+            }
+        }
+
+        if ((is_null($mobilePhoto) || is_null($webPhoto)) && !is_null($photo)) {
+            $mobilePhoto = $photo;
+            $webPhoto = $photo;
+        }
+
+        $name = time() . bin2hex(random_bytes(5));
+
+        if (!is_null($petition['mobile_photo_url']) && ($request->file === 'delete' || $request->file1 === 'delete' || !is_null($mobilePhoto))) {
+            $data['mobile_photo_url'] = null;
+            $mobilePhotoUrl = explode(config('app.server_url'), $petition['mobile_photo_url'])[1];
+            unlink(base_path() . '/storage/app/public/' . $mobilePhotoUrl);
+        }
+        if (!is_null($mobilePhoto)) {
+            $mobilePhotoName = Petition::saveImage($mobilePhoto, Petition::IMAGE_TYPE_MOBILE, $name);
+            $data['mobile_photo_url'] = config('app.server_url') . 'static/' . $mobilePhotoName;
+        }
+
+
+        if (!is_null($petition['web_photo_url']) && ($request->file === 'delete' || $request->file2 === 'delete' || !is_null($webPhoto))) {
+            $data['web_photo_url'] = null;
+            $webPhotoUrl = explode(config('app.server_url'), $petition['web_photo_url'])[1];
+            unlink(base_path() . '/storage/app/public/' . $webPhotoUrl);
+        }
+        if (!is_null($webPhoto)) {
+            $webPhotoName = Petition::saveImage($webPhoto, Petition::IMAGE_TYPE_WEB, $name);
+            $data['web_photo_url'] = config('app.server_url') . 'static/' . $webPhotoName;
+        }
+
+
         Petition::where('id', '=', $petitionId)
             ->update($data);
 
@@ -252,26 +273,25 @@ class PetitionController extends Controller
             return new OkResponse(Petition::getPetitions([$petitionId], $withOwner = true, $friendIds, true, $request->userId, true));
         }
 
-        // TODO: move to consts
         switch ($type) {
-            case 'popular':
+            case Petition::TYPE_POPULAR:
                 return new OkResponse(Petition::getPopular($offset, $friendIds));
 
-            case 'last':
+            case Petition::TYPE_LAST:
                 return new OkResponse(Petition::getLast($offset, $friendIds));
 
-            case 'signed':
+            case Petition::TYPE_SIGNED:
                 return new OkResponse(Petition::getSigned($request->userId, $offset, $friendIds));
 
-            case 'managed':
+            case Petition::TYPE_MANAGED:
                 return new OkResponse(Petition::getManaged($request->userId, $offset, $friendIds));
         }
 
         return new OkResponse([
-            'popular' => Petition::getPopular(0, $friendIds),
-            'last' => Petition::getLast(0, $friendIds),
-            'signed' => Petition::getSigned($request->userId, 0, $friendIds),
-            'managed' => Petition::getManaged($request->userId, 0, $friendIds),
+            Petition::TYPE_POPULAR => Petition::getPopular(0, $friendIds),
+            Petition::TYPE_LAST => Petition::getLast(0, $friendIds),
+            Petition::TYPE_SIGNED => Petition::getSigned($request->userId, 0, $friendIds),
+            Petition::TYPE_MANAGED => Petition::getManaged($request->userId, 0, $friendIds),
         ]);
     }
 }
